+++ packages/typespec-go/test/local/armtest/zz_models.go
@@ -53,6 +53,39 @@ type BodyRootProperties struct {
 	ProvisioningState *ProvisioningState
 }
 
+// Configuration resource.
+type Configuration struct {
+	// REQUIRED; The geo-location where the resource lives
+	Location *string
+
+	// The resource-specific properties for this resource.
+	Properties *ConfigurationProperties
+
+	// Resource tags.
+	Tags map[string]*string
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string
+
+	// READ-ONLY; The name of the resource
+	Name *string
+
+	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
+	SystemData *SystemData
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string
+}
+
+// ConfigurationProperties - Properties of the configuration resource.
+type ConfigurationProperties struct {
+	// Description of the configuration.
+	Description *string
+
+	// READ-ONLY; Status of the current operation on configuration.
+	ProvisioningState *ProvisioningState
+}
+
 // LROModel - LRO model resource.
 type LROModel struct {
 	// REQUIRED; The geo-location where the resource lives