+++ packages/typespec-go/test/local/armtest/zz_options.go
@@ -14,6 +14,23 @@ type BodyRootsClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
+// ConfigurationsClientGetContentOptions contains the optional parameters for the ConfigurationsClient.GetContent method.
+type ConfigurationsClientGetContentOptions struct {
+	// placeholder for future optional parameters
+}
+
+// ConfigurationsClientGetStreamingContentOptions contains the optional parameters for the ConfigurationsClient.GetStreamingContent
+// method.
+type ConfigurationsClientGetStreamingContentOptions struct {
+	// placeholder for future optional parameters
+}
+
+// ConfigurationsClientPutStreamingContentOptions contains the optional parameters for the ConfigurationsClient.PutStreamingContent
+// method.
+type ConfigurationsClientPutStreamingContentOptions struct {
+	// placeholder for future optional parameters
+}
+
 // LROClientBeginOkResponseWithAsyncHeaderOptions contains the optional parameters for the LROClient.BeginOkResponseWithAsyncHeader
 // method.
 type LROClientBeginOkResponseWithAsyncHeaderOptions struct {